Move shell to apps.c
